Move RelatedItemInfo next to fragment using it
